requirements.txt
streamlit
google-generativeai
plotly
python-dotenv
youtube-search